Split local-dev approval gate tests
